Added new function
I didn't find anything wrong with your program so I added an option for the user to display all items of a certain type. It is function 11.  If the user inputs items of the same type into the tree the function will search the tree and display all items of same type along with the year.